### 2020-10-18 #2760 @whedon
[PRE REVIEW]: simstudy: Illuminating research methods through data generation
**Submitting author:** @assignUser (<a href="http://orcid.org/0000-0002-7281-3989">Jacob Wujciak-Jens</a>)
**Repository:** <a href="https://github.com/kgoldfeld/simstudy/" target ="_blank">https://github.com/kgoldfeld/simstudy/</a>
**Version:** v0.2.1
