### 2020-10-19 bc7afec
Updating README.Rmd - adding dist
