+++ README.Rmd
@@ -49,8 +49,8 @@ Here is some simple sample code, much more in the vignettes:
 ```{r, echo = TRUE}
 library(simstudy)
 
-def <- defData(varname="x", formula = 10, variance = 2)
-def <- defData(def, varname="y", formula = "3 + 0.5 * x", variance = 1)
+def <- defData(varname="x", formula = 10, variance = 2, dist = "normal")
+def <- defData(def, varname="y", formula = "3 + 0.5 * x", variance = 1, dist = "normal")
 dd <- genData(250, def)
 
 dd <- trtAssign(dd, nTrt = 4, grpName = "grp", balanced = TRUE)
