+++ dev/pkgdown.yml
@@ -12,5 +12,5 @@ articles:
   spline: spline.html
   survival: survival.html
   treat_and_exposure: treat_and_exposure.html
-last_built: 2020-10-08T22:39Z
+last_built: 2020-10-19T14:12Z
 
