### 2020-10-20 #2763 @brunaw
I  personally appreciate the creation of this package, since I have had to come up with my own data simulation procedures several times, for many different contexts. I think this package should be accepted because it will help lots of researchers with the same issues. Comments and suggestions about the code & content/documentation of the package are below.

## Package content/documentation

- Add latex to the README for clarity
- There are a few grammatical errors in the vignettes that could be improved, I suggest using Grammarly for a text review.

- I don't think the description for the `formula` and `variance` arguments is clear yet. For instance, what does it mean to say that `formula = 10`? Likewise, what is the `variance` argument related to? I think a good way of solving this would be a concrete, written-up example with a Normal distribution, showing that the`variance` argument is the variance of the Normal distribution you're simulating from, and what does the `formula` argument represent in that.

- It might be useful to use a simulation seed (`set.seed()`) throughout the vignettes, so people will know that their code is working as it should be when replicating it. For some vignettes, only running the code leads to quite different results from what is seen in the package website (since it's all random).

- I think it would be nice to have your references listed in the vignettes. For example, simulating correlated (multivariate data) is not a trivial thing, since we don't have as much theory available for multivariate distributions beyond the Normal case. I had to go check the reference you cite in the JOSS paper for the binary case, since I hadn't come across that method before. Other users might be interested in knowing exactly what is going on in the code, especially if they need to explain that later (in a paper, for example).

- You state the existence of the Contributor Code of Conduct, but actual Contributor guidelines are not provided, or how to report a problem. I think it would be good to write a few lines in the README addressing that.

## Package code
- The package installs locally (with a cloned repository and CRAN) but it wouldn't install from GitHub, the following URL was not found:
> https://api.github.com/repos/kgoldfeld/simstudy/tarball/master

- The package builds okay after installing the packages used in the vignettes
- Tests are provided and also run correctly
- I haven't found any other issues with the code itself
