### 2020-10-20 #2763 @assignUser
@brunaw Thank you for your review and your positive recommendation! I will try to address each point you have brought up

* Could you elaborate in what way you would like latex to be added to the README?
* We will review the texts.
* The meaning of each parameter for all available distributions is explained in the package help under [`?distributions`](https://kgoldfeld.github.io/simstudy/reference/distributions.html) and in the overview [vignette](https://kgoldfeld.github.io/simstudy/articles/simstudy.html). Is this in line with what you had in mind?
* I agree on the seed, which is already the case in some vignettes, we will work on consistency there. Or are there reasons not to @kgoldfeld?
* It makes sense to add the reference(s) to the documentation.
* CoC and Contributor Guidlines are in `.github/CODE_OF_CONDUCT.md` and  `.github/CONTRIBUTING.md` respectively so they should be picked up by github for new users, furthermore  they are also prominently linked in the [pkgdown site](https://kgoldfeld.github.io/simstudy/). Maybe we could add a note to the README about contributions/error reports @kgoldfeld?
* How are you installing from github? There seems to be a mention of the `master` branch which we renamed to `main`. I tested `devtools::install_github("kgoldfeld/simstudy")` and that works as intended. I just noticed & fixed that the codemeta (generated with `codemetar`) was referring to the the master branch for news as well.
