### 2020-10-20 #2763 @brunaw
@assignUser Thank you for the quick reply, and I apologize if some of my comments weren't clear.

- For example, you have log(\mu)  in the text that could be shown as <a href="https://www.codecogs.com/eqnedit.php?latex=log(\mu)" target="_blank"><img src="https://latex.codecogs.com/gif.latex?log(\mu)" title="log(\mu)" /></a>. This can be quite tricky (https://stackoverflow.com/questions/35498525/latex-rendering-in-readme-md-on-github) to do and not necessarily pretty, but I think it could be useful for you (it's only a suggestion though)
- Okay, thank you
- Yes, exactly, I thought that even with the explanation it wasn't clear for the general user. People who do simulation frequently might get it when reading it, but it would be helpful if the argument definitions were explained in a bit more detail.
- Okay, thank you
- Okay, thank you
-  Okay, I might have missed it on the pkgdown website, sorry about that! Maybe just link it to the README as well
-  I wish I could come up with a more helpful explanation for you, but I couldn't wrap my head around it either. I just run `devtools::install_github("kgoldfeld/simstudy")` and get the error. I thought it was an internet connection problem but I tried it on a different network (just now) and the same error happens. The code doesn't suggest anything about it either, so I don't know what causes it.
