+++ README.md
@@ -72,6 +72,7 @@ Here is some simple sample code, much more in the vignettes:
 
 ``` r
 library(simstudy)
+set.seed(1965)
 
 def <- defData(varname="x", formula = 10, variance = 2, dist = "normal")
 def <- defData(def, varname="y", formula = "3 + 0.5 * x", variance = 1, dist = "normal")
@@ -81,17 +82,17 @@ dd <- trtAssign(dd, nTrt = 4, grpName = "grp", balanced = TRUE)
 
 dd
 #>       id         x        y grp
-#>   1:   1 10.343550 7.896122   1
-#>   2:   2 12.546128 8.213661   2
-#>   3:   3 11.589142 9.335444   4
-#>   4:   4  6.780056 6.561719   3
-#>   5:   5 10.826578 7.601935   1
+#>   1:   1 11.191960 8.949389   4
+#>   2:   2 10.418375 7.372060   2
+#>   3:   3  8.512109 6.925844   4
+#>   4:   4 11.361632 9.850340   4
+#>   5:   5  9.928811 6.515463   2
 #>  ---                           
-#> 246: 246 10.050552 6.066714   3
-#> 247: 247 11.833109 7.482641   2
-#> 248: 248 10.065499 8.740391   2
-#> 249: 249  9.853735 6.735114   1
-#> 250: 250  8.823834 5.438522   2
+#> 246: 246  8.220609 7.898416   4
+#> 247: 247  8.531483 8.681783   4
+#> 248: 248 10.507370 8.552350   4
+#> 249: 249  8.621339 6.652300   1
+#> 250: 250  9.508164 7.083845   4
 ```
 
 ## Code of Conduct
