### 2020-10-20 d69542a
Proper latex for README.Rmd
