### 2020-10-20 07fc50c
Fixing README.md latex
