### 2020-10-20 #2763 @assignUser
@brunaw
Latex: Ah I see. As the symbols are rendered correctly on the pkgdown page and any workarounds for github might interfere with that I think we will keep it as is (sadly, I would like it to be rendered too, maybe github will add it at some point).

Arguments: I feel that with all of the vignettes and targeted documentation we have this covered as a specific example for `normal` would not cover all use cases, unlike the distributions. Please let me know if this is still a sticking point for you.

Install: I just found the solution for this issue. This is a recently solved problem where devtools/remotes was assuming the default branch to be "master" which was fixed by https://github.com/r-lib/remotes/pull/510 but is not on cran yet: https://remotes.r-lib.org/news/index.html
You should be able to install by setting `devtools::install_github("kgoldfeld/simstudy", ref = "HEAD")`
