+++ README.md
@@ -46,9 +46,9 @@ generated based on the specification of a mean μ (or log(μ)) and a
 dispersion \(d\), rather than shape α and rate β parameters that more
 typically characterize the *gamma* distribution. When we estimate the
 parameters, we are modeling μ (or some function of μ), so we should
-explicitly recover the `simstudy` parameters used to generate the model
-- illuminating the relationship between the underlying data generating
-processes and the models.
+explicitly recover the `simstudy` parameters used to generate the model,
+thus illuminating the relationship between the underlying data
+generating processes and the models.
 
 ## Installation
 
