### 2020-10-20 3f43836
Adding contributing instructions to README.md
