+++ README.Rmd
@@ -59,6 +59,10 @@ dd <- trtAssign(dd, nTrt = 4, grpName = "grp", balanced = TRUE)
 dd
 ```
 
+## Contributing & Support
+
+If you find a bug or need help, please file an issue with a [reprex](https://www.tidyverse.org/help/) on [Github](https://github.com/kgoldfeld/simstudy/issues). We are happy to accept contributions to simstudy. More information on how to propose changes or fix bugs can be found [here](https://kgoldfeld.github.io/simstudy/CONTRIBUTING.html).
+
 ## Code of Conduct
 
 Please note that the simstudy project is released with a [Contributor Code of Conduct](https://kgoldfeld.github.io/simstudy/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
