### 2020-10-20 #2763 @kgoldfeld
@brunaw
I was able to fix the LaTex issue without affecting the pkgdown site. Thanks for that suggestion - it looks much better - it always bothered me.

As for the grammar of the vignettes - were there any that were particularly egregious? I will certainly go through all of them - as you see there are quite a few - but it would help if you found that particular ones need special attention.
