### 2020-10-20 644b87e
Updating with edited vignettes
