+++ vignettes/longitudinal.Rmd
@@ -46,11 +46,13 @@ tdef <- defData(tdef, varname = "Y0", dist = "normal", formula = 10, variance = 
 tdef <- defData(tdef, varname = "Y1", dist = "normal", formula = "Y0 + 5 + 5 * T", variance = 1)
 tdef <- defData(tdef, varname = "Y2", dist = "normal", formula = "Y0 + 10 + 5 * T", variance = 1)
 
+set.seed (483726)
+
 dtTrial <- genData( 500, tdef)
 dtTrial
 ```
 
-The data in longitudinal form is created with a call to **`addPeriods`**. If the cross-sectional data includes time dependent data, then the number of periods `nPeriods` must be the same as the number of time dependent columns. If a variable is not declared as one of the `timevars`, it will be repeated each time period. In this example, the treatment indicator `T` is not specified as a time dependent variable. (Note: if there are two time-dependent variables, it is best to create two data sets and merge them. This will be shown later in the vignette).
+Longitudinal data are created with a call to **`addPeriods`**. If the cross-sectional data includes time-dependent data, then the number of periods `nPeriods` must be the same as the number of time-dependent columns. If a variable is not declared as one of the `timevars`, it will be repeated each time period. In this example, the treatment indicator `T` is not specified as a time-dependent variable. (Note: if there are two time-dependent variables, it is best to create two data sets and merge them. This will be shown later in the vignette).
 
 ```{r, tidy = TRUE}
 dtTime <- addPeriods(dtTrial, nPeriods = 3, idvars = "id", timevars = c("Y0", "Y1", "Y2"), timevarName = "Y")
@@ -76,7 +78,7 @@ ggplot(data = dtTime, aes(x = factor(period), y = Y)) +
 
 ### Longitudinal data with varying observation and interval times
 
-It is also possible to generate longitudinal data with varying numbers of measurement periods as well as varying time intervals between each measurement period. This is done by defining specific variables in the data set that define the number of observations per subject and the average interval time between each observation. `nCount` defines the number of measurements for an individual; `mInterval` specifies the average time between intervals for an subject; and `vInterval` specifies the variance of those interval times. If `vInterval` is set to 0 or is not defined, the interval for a subject is determined entirely by the mean interval. If `vInterval` is greater than 0, time intervals are generated using a gamma distribution with mean and dispersion specified.
+It is also possible to generate longitudinal data with varying numbers of measurement periods as well as varying time intervals between each measurement period. This is done by defining specific variables in the data set that define the number of observations per subject and the average interval time between each observation. `nCount` defines the number of measurements for an individual; `mInterval` specifies the average time between intervals for a subject; and `vInterval` specifies the variance of those interval times. If `vInterval` is set to 0 or is not defined, the interval for a subject is determined entirely by the mean interval. If `vInterval` is greater than 0, time intervals are generated using a gamma distribution with mean and dispersion specified.
 
 In this simple example, the cross-sectional data generates individuals with a different number of measurement observations and different times between each observation. Data for two of these individuals is printed:
 
@@ -97,14 +99,14 @@ dtPeriod <- addPeriods(dt)
 dtPeriod[id %in% c(8,121)]  # View individuals 8 and 121 only
 ```
 
-If a time sensitive measurement is added to the data set ...
+If a time-sensitive measurement is added to the data set ...
 
 ```{r, tidy = TRUE}
 def2 <- defDataAdd(varname = "Y", dist = "normal", formula = "15 + .1 * time", variance = 5)
 dtPeriod <- addColumns(def2, dtPeriod)
 ```
 
-... a plot of a five randomly selected individuals looks like this:
+... a plot of five randomly selected individuals looks like this:
 
 ```{r, tidy = TRUE, echo = FALSE, fig.width = 6, fig.height = 3}
 
