+++ vignettes/simstudy.Rmd
@@ -15,8 +15,6 @@ library(ggplot2)
 library(knitr)
 library(data.table)
 
-set.seed(33333)
-
 options(digits = 3)
 
 opts_chunk$set(tidy.opts=list(width.cutoff=55), tidy=TRUE)
@@ -42,7 +40,7 @@ ggtheme <- function(panelback = "white") {
 }
 ```
 
-The `simstudy` package is collection of functions that allow users to generate simulated data sets in order to explore modeling techniques or better understand data generating processes. The user defines the distributions of individual variables, specifies relationships between covariates and outcomes, and generates data based on these specifications. The final data sets can represent randomized control trials, repeated measure designs, cluster randomized trials, or naturally observed data processes. Other complexities that can be added include survival data, correlated data, factorial study designs, step wedge designs, and missing data processes.
+The `simstudy` package is a collection of functions that allows users to generate simulated data sets to explore modeling techniques or better understand data generating processes. The user defines the distributions of individual variables, specifies relationships between covariates and outcomes, and generates data based on these specifications. The final data sets can represent randomized control trials, repeated measure designs, cluster-randomized trials, or naturally observed data processes. Other complexities that can be added include survival data, correlated data, factorial study designs, step wedge designs, and missing data processes.
 
 Simulation using `simstudy` has two fundamental steps. The user (1) **defines** the data elements of a data set and (2) **generates** the data based on these definitions. Additional functionality exists to simulate observed or randomized **treatment assignment/exposures**, to create **longitudinal/panel** data, to create **multi-level/hierarchical** data, to create datasets with **correlated variables** based on a specified covariance structure, to **merge** datasets, to create data sets with **missing** data, and to create non-linear relationships with underlying **spline** curves.
 
@@ -54,7 +52,7 @@ This introduction provides a brief overview to the basics of defining and genera
 
 ### Defining the Data
 
-The key to simulating data in `simstudy` is the creation of series of data definition tables that look like this:
+The key to simulating data in `simstudy` is the creation of a series of data definition tables that look like this:
 
 ```{r,  echo=FALSE}
 def <- defData(varname="age", dist="normal", formula=10, variance = 2)
@@ -66,7 +64,7 @@ def <- defData(def,varname="visits", dist="poisson",
 knitr::kable(def)
 ```
 
-These definition tables can be generated two ways. One option is to to use any external editor that allows the creation of .csv files, which can be read in with a call to `defRead`. An  alternative is to make repeated calls to the function `defData`. This script builds a definition table internally:
+These definition tables can be generated in two ways. One option is to use any external editor that allows the creation of .csv files, which can be read in with a call to `defRead`. An alternative is to make repeated calls to the function `defData`. This script builds a definition table internally:
 
 ```{r}
 def <- defData(varname="age", dist="normal", formula=10, variance = 2)
@@ -76,7 +74,7 @@ def <- defData(def,varname="visits", dist="poisson",
   formula="1.5 - 0.2 * age + 0.5 * female", link="log")
 ```
 
-The data definition table includes a row for each variable that is to be generated, and has the following fields: **varname**, **formula**, **variance**, **dist**, and **link**. **varname** is a provides the name of the variable to be generated. **formula** is either a value or string representing any valid R formula (which can include function calls) that in most cases defines the mean of the distribution. **variance** is a value or string that specifies either the variance or other parameter that characterizes the distribution; the default is 0. **dist** is defines the distribution of the variable to be generated; the default is *normal*. The **link** is a function that defines the relationship of the formula with the mean value, and can either *identity*, *log*, or *logit*, depending on the distribution; the default is *identity*.
[…]

[…]
