### 2020-10-20 #2763 @kgoldfeld
@brunaw
I was able to go through the vignettes - there were some really bad typos throughout. Thanks so much for taking the time to catch all that. I have added seeds as well throughout where there were none. Thanks again.
