### 2020-10-21 #2763 @gagolews
The package may be useful to some researchers and students,
in cases where generation of data following some typical models
is required.
Overall, it is quite well written and well documented.
The package's API can be considered user friendly.
Yet, of course, more exotic scenarios will require its users to
implement the missing functionality by hand anyway (and learn how
to implement the models included in `simstudy` anyway).
I recommended the paper be accepted provided that the authors
address what follows.

# Paper — remarks:

1. Neither in the title nor in the summary the reader is informed that it is an **R** package

2. Mention it's availability on CRAN, add link to CRAN entry, and info on how to install the package

3. The package is documented quite well (vignettes),
the link to https://kgoldfeld.github.io/simstudy/articles/simstudy.html should
therefore be emphasised in the main text,
as in *"for more details on the package, use cases, etc. (...), see (...)"*.
Also, add this information to the README file.

4. `$log(\mu)$` → `$\log(\mu)$`

# Vignettes — remarks:

1. https://kgoldfeld.github.io/simstudy/articles/simstudy.html

    a. `$log(\mu)$` → `$\log(\mu)$`

    b. "please refer to other package vignettes" – you mean other vignettes
    included in this very package?

    c. "One option is to to use"

    d. "has the following fields: varname, formula, variance, dist, and link" —
    consider using `varname`, `formula` (code)

    e. e.g., in `defData(def, varname = "female", dist = "binary",
    formula = "-2 + age * 0.1", link = "logit")` — I guess a more R-way
    (which is a matter of taste) would be to specify
    `formula = -2+age*0.1`, i.e., as an R expression, see `?deparse`
    and `?substitute`. The same with `varname` and `dist`.
    See base R functions `transform()` and `subset()` for inspiration.

    f. `$uniform$` → `uniform`

    g. `$uniforminteger$` → `uniforminteger`

2. https://kgoldfeld.github.io/simstudy/articles/correlated.html

    a. side note: copulas (copulae) are nice tools for modelling
    of dependencies between random variables
