### 2020-10-21 #2763 @gagolews
@brunaw re: simulating correlated data - see https://en.wikipedia.org/wiki/Copula_(probability_theory) and https://cran.r-project.org/web/packages/copula/index.html
