### 2020-10-21 #2763 @kgoldfeld
@gagolews Thanks for your feedback - will make the editing changes. I agree with you regarding 1e - we may make that change in a future iteration, though there are some "formulas" where there is no standard R formulation. The following (clearly nonsensical) snippet shows three distributions where the standard formulas don't really apply:

```r
library(simstudy)

d <- defData(varname = "x", formula = "0;1", dist = "uniform")
d <- defData(d, varname ="y", formula = "-2+x;-1 + 0.5*x",
             dist = "categorical", link = "logit")
d <- defData(d, varname = "z", formula = "x|0.5 + y|0.5", dist = "mixture")

set.seed(5)
genData(10, d)
#>     id         x y         z
#>  1:  1 0.2002145 3 0.2002145
#>  2:  2 0.6852186 3 3.0000000
#>  3:  3 0.9168758 3 0.9168758
#>  4:  4 0.2843995 2 0.2843995
#>  5:  5 0.1046501 2 0.1046501
#>  6:  6 0.7010575 1 0.7010575
#>  7:  7 0.5279600 3 0.5279600
#>  8:  8 0.8079352 3 0.8079352
#>  9:  9 0.9565001 3 3.0000000
#> 10: 10 0.1104530 1 1.0000000
```

With respect to 2a, I agree that copulas are useful in generating correlated data. Indeed, that is what we are doing in `simstudy` to generated correlated data (for all distributions other than the normal distribution).
