### 2020-10-21 b31941a
Updating simstudy.Rmd
