### 2020-10-21 #2763 @gagolews
Yeah, but this way you end up inventing a totally new syntax, like a "language within a language", making the API more difficult to learn (and these skills are "not transferable" to other package, as I understand the package aims at less advanced users?)

Actually `x|0.5 + y|0.5` is a valid R expression although it means `x|(0.5+y)|0.5` and some users might read it exactly like this?
How about simply `0.5*x + 0.5*y` for mixtures? By the way in `lm.formula()` we also use the `I(...)` function to transform the variables

`"0;1"` → `c(0,1)` ?
`"-2+x;-1 + 0.5*x"` → `c(...)?

I reckon these are/were all pretty difficult API design questions. If I were you, I'd opt for maximum compatibility with other R functions, especially those from base R, `stats`, etc. But that's my 99 cents, it's a matter of taste.
