### 2020-10-21 #2763 @kgoldfeld
I agree - and it was definitely something I struggled with.  There are some things I wish I had done differently, like naming `formula` and `variance` rather than something like `param1` and `param2`. We may move to that at some point, and could think about the specification as well.

As for who is using it, I think it is a mix of experienced and less experienced folks. I think the experienced folks appreciate the ease with which you can generate complex study designs without a ton of coding.

I am hesitant to make major changes without some serious thought, because there have been 30K+ downloads, which probably means a few hundred or so serious users. I don't want to mess them up. But I hear what you are saying.
