### 2020-10-21 #2763 @kgoldfeld
I made all the editing changes to the  vignette - my collaborator will fix the paper issues tomorrow at some point.
