+++ paper/simstudy.bib
@@ -205,6 +205,16 @@
   keywords = {⛔ No DOI found}
 }
 
+@manual{rcoreteam2020,
+  title = {R: {{A}} Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  date = {2020},
+  location = {{Vienna, Austria}},
+  url = {https://www.R-project.org/},
+  organization = {{R Foundation for Statistical Computing}},
+  type = {manual}
+}
+
 @article{renson2017,
   title = {Lack of Insurance Is Associated with Lower Probability of Diagnostic Imaging Use among {{US}} Trauma Patients: {{An}} Instrumental Variable Analysis and Simulation},
   author = {Renson, Audrey and Schubert, Finn D and Bjurlin, Marc A},
