### 2020-10-21 cdd430a
fix math
