### 2020-10-21 #2763 @assignUser
@gagolews Thank you for your review, positive recommendation  and valuable insights. I have opened an issue regarding 1.e kgoldfeld/simstudy/issues/75 to keep track of it when we are considering breaking changes.

I have made the changes to the paper and readme.
