### 2020-10-21 #2763 @brunaw
@assignUser and @kgoldfeld  Thank you for your answers. The package correctly installs from GitHub now, and as for the grammar, I didn't mean anything specific but a few typos here and there that could be fixed (which you already did). In these cases, I always use Grammarly because I'm sure I'll miss something.  My other issue was mostly regarding this line of code

```
def <- defData(varname = "age", dist = "normal", formula = 10,
               variance = 2)
genData(5, def)
```

where you're simulating from a N(10, 2) distribution. When I read the documentation I think I missed the part where you say that the formula argument represents the mean of the distribution, so this was actually my mistake, sorry about that! No need to correct anything there.

@gagolews Copulas are the most popular way of simulating from correlated data, but they're also limited because they depend on assumptions about the underlying multivariate distribution and about the marginal distributions. For Binomial cases,  for instance, this is not very well defined in the literature as for a Normal distribution (e.g. https://brunaw.com/slides/conferences/EMR2019_poster.pdf).  As a result, you have a set of possible methods to simulate from non-Gaussian correlated data, but most times they present issues such as not allowing full flexibility of the covariance matrix. That's why this is not a trivial thing to do deal with, and why I suggested some references are given in the vignettes.
