### 2020-10-21 #2763 @assignUser
@brunaw Thanks again :) If you are satisfied with the new "Contributing & Support"  section of the readme pointing to CONTRIBUTING.md could you check your last box? :)
