### 2020-10-21 75b5d8d
Adding copula links to vignette
