### 2020-10-21 #2763 @kgoldfeld
@assignUser @brunaw I just updated the correlation vignette to include the two links that Bruna recommended.

Thanks again for all the comments and suggestions.
