### 2020-10-26 #2763 @mikldk
@gagolews, @brunaw:  Can you confirm that you have finished the review and recommend that this paper is now published?

@assignUser:

* Please make a tagged release and archive (e.g. with [Zenodo](https://zenodo.org/)) as described [here](<https://joss.readthedocs.io/en/latest/submitting.html#the-review-process>), and report the version number and archive DOI in this thread. Please verify that the archive deposit has the correct metadata (title and author list), or edit these if that is not the case.
