+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Type: Package
 Package: simstudy
 Title: Simulation of Study Data
-Version: 0.2.1
-Date: 2020-10-07
+Version: 0.2.2
+Date: 2020-10-26
 Authors@R: 
     c(person(given = "Keith",
              family = "Goldfeld",
