### 2020-10-26 b2d3382
Fixing typos for publishing
