### 2020-10-26 #2763 @assignUser
@mikldk @kgoldfeld The tagged release is https://github.com/kgoldfeld/simstudy/releases/tag/v0.2.2 archived on Zenodo with the DOI: https://doi.org/10.5281/zenodo.4134675
meta data should be correct.
